Move MKV Bluray parsing from BinEdit.cs into libse for use by third-party code.
